Backend/requirements.txt
flask
flask-cors
flask-socketio
python-dotenv
pymongo
eventlet
